requirements.txt
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
fonttools==4.58.4
idna==3.10
imbalanced-learn==0.13.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
pyaml==25.5.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
xgboost==3.0.2
